Restructure project as monorepo with separate @mkja/flip and @mkja/flip-group packages
