+++ README.md
@@ -1,84 +1,22 @@
-# flip-motion
+# flip.js monorepo
 
-Tiny FLIP animation utility (translate + optional scale) with Promise-based `play` and reduced-motion support.
+This repo contains two ESM-only packages:
 
-## Install
+- `@mkja/flip`: core FLIP utilities (imperative)
+- `@mkja/flip-group`: a declarative custom element that uses `@mkja/flip`
 
-Install from npm:
+Both ship types generated from JSDoc.
 
-```bash
-npm install @mkja/flip-motion
-```
+## Workspace commands
 
-## Usage
+- Build types: `npm run build`
+- Test core: `npm test -w @mkja/flip`
+- Publish: `npm publish --access public -w <package>`
 
-```js
-import flip from '@mkja/flip-motion';
-
-const elements = document.querySelectorAll('li');
-const ctrl = flip(elements);
-
-// Mutate DOM (reorder/insert/remove)
-// ...
-
-await ctrl.play({ duration: 300, easing: 'ease' }).finished;
-```
-
-## API
-
-- `flip(elements)` → controller with:
-  - `play(options)` → `{ animations: Animation[], finished: Promise<void>, cancel() }`
-  - `measure()` → refresh internal measurements
-  - `update([elements])` → rebind or just refresh measurements
-  - `cancel()` → cancel in-flight animations
-  - `disconnect()` → release references
-
-### Options
-
-- `duration` (ms, default 100)
-- `easing` (default `ease`). Accepts any CSS timing function string supported by WAAPI: `linear`, `ease`, `ease-in`, `ease-out`, `ease-in-out`, `step-start`, `step-end`, `cubic-bezier(...)`, `steps(n[, start|end])`, `linear(...)`.
-- `delay` (default 0)
-- `stagger` (ms per index, default 0)
-- `fill` (default `both`)
-- `direction` (default `normal`)
-- `composite` (default `add`)
-- `shouldScale` (default `true`)
-- `respectReducedMotion` (default `true`)
-- `transformOrigin` (default `"0 0"`)
-- `epsilon` (px threshold to skip, default `0.5`)
-
-### Per-element overrides (data attributes)
-
-- `data-flip-duration`: absolute duration in ms for that element (overrides `options.duration`)
-- `data-flip-duration-offset`: value added to `options.duration` for that element (ignored if `data-flip-duration` is present)
-- `data-flip-delay`: absolute delay in ms for that element (overrides base `delay` + `stagger`)
-
-Example:
-
-```html
-<li data-flip-duration="200"></li>
-<li data-flip-duration-offset="50" data-flip-delay="100"></li>
-```
-
-Or via JS:
+## Import
 
 ```js
-el.dataset.flipDuration = '200';
-el.dataset.flipDurationOffset = '50';
-el.dataset.flipDelay = '100';
+import { flip } from '@mkja/flip';
+// or custom element
+import '@mkja/flip-group';
 ```
-
-## Development
-
-- Dev demo: `npm run dev` (serves `test/index.html` with Vite)
-- Tests: `npm test`
-- Lint: `npm run lint`
-- Format: `npm run format`
-
-## Browser support
-
-Uses WAAPI (`element.animate`). If unavailable, calls are no-ops. Honors `prefers-reduced-motion`.
-
-## License
-
-MIT
