--- eslint.config.js
@@ -1,21 +1,0 @@
-import js from '@eslint/js';
-import globals from 'globals';
-
-export default [
-  {
-    ignores: ['dist/**', 'coverage/**', 'node_modules/**', 'test/**'],
-  },
-  {
-    languageOptions: {
-      ecmaVersion: 'latest',
-      sourceType: 'module',
-      globals: {
-        ...globals.browser,
-        ...globals.node,
-      },
-    },
-    rules: {
-      ...js.configs.recommended.rules,
-    },
-  },
-];
