--- jsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "checkJs": true,
-    "target": "ES2020",
-    "moduleResolution": "bundler"
-  },
-  "exclude": ["node_modules"]
-}
