+++ packages/flip-group/README.md
@@ -1,0 +1,19 @@
+# @mkja/flip-group
+
+Declarative FLIP custom element built on `@mkja/flip`.
+
+## Install
+
+```bash
+npm i @mkja/flip-group @mkja/flip
+```
+
+## Use
+
+```js
+import '@mkja/flip-group';
+```
+
+```html
+<flip-group selector="li" stagger="proximity" duration="300"></flip-group>
+```
