+++ packages/flip-group/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@mkja/flip-group",
+  "version": "0.0.0",
+  "description": "Declarative DOM FLIP helper custom element",
+  "license": "MIT",
+  "type": "module",
+  "peerDependencies": {
+    "@mkja/flip": "^0.0.0"
+  },
+  "devDependencies": {
+    "vite": "^5.4.2",
+    "vitest": "^2.0.5"
+  },
+  "exports": {
+    ".": "./src/index.js"
+  },
+  "types": "./types/index.d.ts",
+  "files": ["src", "types", "README.md", "LICENSE"],
+  "sideEffects": false,
+  "scripts": {
+    "types": "tsc -p tsconfig.types.json",
+    "build": "npm run types",
+    "test": "vitest --run"
+  },
+  "engines": {
+    "node": ">=18.0.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/m-kj-aldrin/flip.js.git",
+    "directory": "packages/flip-group"
+  }
+}
