+++ packages/flip-group/src/index.js
@@ -1,0 +1,1 @@
+export { FlipGroup } from './flip-group.js';
