+++ packages/flip-group/tsconfig.types.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "./types"
+  },
+  "include": ["src/index.js", "src/flip-group.js", "src/flip-group.d.ts"],
+  "exclude": ["**/*.test.js", "**/__tests__/**"]
+}
