+++ packages/flip-group/types/index.d.ts
@@ -1,0 +1,1 @@
+export { FlipGroup } from "./flip-group.js";
