+++ packages/flip/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@mkja/flip",
+  "version": "0.0.0",
+  "description": "Core FLIP animation utilities",
+  "license": "MIT",
+  "type": "module",
+  "exports": {
+    ".": "./src/index.js"
+  },
+  "types": "./types/index.d.ts",
+  "files": ["src", "types", "README.md", "LICENSE"],
+  "sideEffects": false,
+  "scripts": {
+    "types": "tsc -p tsconfig.types.json",
+    "build": "npm run types",
+    "test": "vitest --run"
+  },
+  "devDependencies": {
+    "vite": "^5.4.2",
+    "vitest": "^2.0.5"
+  },
+  "engines": {
+    "node": ">=18.0.0"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/m-kj-aldrin/flip.js.git",
+    "directory": "packages/flip"
+  }
+}
