+++ packages/flip/src/index.js
@@ -1,0 +1,1 @@
+export { default as flip } from './flip.js';
