+++ packages/flip/types/index.d.ts
@@ -1,0 +1,1 @@
+export { default as flip } from "./flip.js";
