+++ tsconfig.base.json
@@ -1,0 +1,13 @@
+{
+  "compilerOptions": {
+    "allowJs": true,
+    "checkJs": true,
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "declarationMap": false,
+    "skipLibCheck": true,
+    "noEmit": false,
+    "target": "ES2020",
+    "moduleResolution": "bundler"
+  }
+}
