--- tsconfig.types.json
@@ -1,15 +1,0 @@
-{
-  "extends": "./jsconfig.json",
-  "compilerOptions": {
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "noEmit": false,
-    "outDir": "types",
-    "allowJs": true,
-    "checkJs": true,
-    "skipLibCheck": true
-  },
-  "include": ["src/flip.js"]
-}
-
-
