--- types/index.d.ts
@@ -1,2 +1,0 @@
-/// <reference path="./flip.d.ts" />
-/// <reference path="./dom-augmentations.d.ts" />
