+++ vitest.config.js
@@ -2,8 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    environment: 'jsdom',
-    include: ['__tests__/**/*.{js,ts}'],
+    include: ['packages/*/{src,__tests__}/**/*.test.{js,ts}'],
     coverage: {
       reporter: ['text', 'lcov'],
     },
